Pokémon Go!
Pokémon Go! became a very famous augmented reality (AR) game in 2016
summer. In this project, we want to understand the success of the mobile
app game. Specifically, the purposes of this project are do web scraping
using BeautifulSoup4, to construct a Pandas data frame, to
explore/visualize the numeric data using matplotlib or seaborn, analyze
the app’s screenshot images using deep learning with tensorflow and
finally to use sklearn to build machine learning models to predict the
app’s review counts.